feat: 테스트 분리 및 서블릿 매핑 테스트 추가
